moved not found page to component
