+++ src/components/Layout/Layout.jsx
@@ -1,15 +1,16 @@
 import { Suspense } from 'react';
 import { Outlet } from 'react-router-dom';
 import { Header } from 'components/Header/Header';
-import { Footer } from 'components/Footer/Footer'
+import { Footer } from 'components/Footer/Footer';
+import { Loader } from 'components/Loader/Loader';
 
 export const Layout = () => {
   return (
     <>
-      <Suspense>
-      <Header/>
+      <Suspense fallback={<Loader />}>
+        <Header />
         <Outlet />
-      <Footer/>
+        <Footer />
       </Suspense>
     </>
   );
